# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doctor API",
        "description": "API for managing doctors",
        "contact": {
            "email": "murtada199815@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/AddSpecial": {
            "post": {
                "tags": [
                    "Specialization"
                ],
                "summary": "add Special",
                "description": "Register Special ",
                "operationId": "add-Special",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "special"
                                ],
                                "properties": {
                                    "special": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "special created successfully",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer"
                                        },
                                        "patient": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "No Data"
                    }
                }
            }
        },
        "/api/getAllSpecials": {
            "get": {
                "tags": [
                    "Specialization"
                ],
                "summary": "get Special",
                "description": "get Special ",
                "operationId": "get-Special",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "": {
                                        "type": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successfully",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer"
                                        },
                                        "patient": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Data Not Found!"
                    }
                }
            }
        },
        "/api/updateSpecial/{id}": {
            "put": {
                "tags": [
                    "Specialization"
                ],
                "summary": "update Special",
                "description": "update  Special  Here",
                "operationId": "updateSpecial",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "special": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Special updated successfully",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer"
                                        },
                                        "special": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Data Not Found!"
                    }
                }
            }
        },
        "/api/deleteSpecial/{id}": {
            "delete": {
                "tags": [
                    "Specialization"
                ],
                "summary": "Delete Special",
                "description": "delete Special here",
                "operationId": "deleteSpecial",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application-json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "301": {
                        "description": "Special deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Data Not Found!"
                    }
                }
            }
        }
    },
...
}